feed associado ao login e reorganizaçao de arquivos
