+++ src/router.js
@@ -3,6 +3,7 @@ import { onNavigate } from './utils/history.js';
 import { Feed } from './pages/feed/index.js';
 import { Home } from './home.js';
 import { logOut } from './pages/feed/feed.js';
+import { emailLogin } from './pages/login/login.js';
 
   const routeRender = () => {
   const rootDiv = document.getElementById('root');
@@ -17,12 +18,21 @@ import { logOut } from './pages/feed/feed.js';
 window.addEventListener('popstate', routeRender);
 window.addEventListener('load', () => {
       onNavigate('/')
+      document.getElementById('feedIn').addEventListener("click", (e) =>{
+        e.preventDefault()
+        emailLogin()
+        document.getElementById('boxLogin').style.display = "none";
+      })
       firebase.auth().onAuthStateChanged(function(user) {
         if (user) { 
+          document.getElementById('inicial-page').style.display = "none";
+          document.getElementById('root').style.width = "100%"
           onNavigate('/feed');
-          logOut();
-          
+          logOut();      
        } 
+       else {
+         document.getElementById('inicial-page').style.display = "block";
+       }
       }) 
   routeRender();
 });
